+++ HISTORY.rst
@@ -16,6 +16,14 @@ Unreleased
 * ALCN-957 Add field suite name on partners
 * ALCN-1031: Add narcotic filter on sale report
 * ALCN-1019: Fix mapping for country in Pharmacy export
+* ALCN-1006: Import vet_subscription_number from AS400
+
+**Features and Improvements**
+
+* ALCN-1006: Add field vet_subscription_number
+* ALCN-1006: rename field depot_number to vet_depot_number
+* ALCN-1006: rename field depot_number_visible to is_veterinary
+
 
 **Bugfixes**